refactor: consolidate database utilities and refine S3 large file uploads
- Move 'Database' utility object from 'database.ts' to the main database index
- Remove unused 'MONITOR_SECRET' from '.env.example' and 'README'
- Improve S3 multipart upload by explicitly calling 'writer.flush()' after each chunk
- Clean up redundant file 'app/database/database.ts'
