+++ .env.example
@@ -1,6 +1,5 @@
 # Application
 APP_ID=
-MONITOR_SECRET=
 ENV=development
 PORT=3000
 MODE=fullstack
